+++ packages/expressive-code/CHANGELOG.md
@@ -1,5 +1,7 @@
 # expressive-code
 
+## 0.7.0
+
 ## 0.6.0
 
 ### Patch Changes